spider function into .cc file
The body functions should probably be in
their own cc files anyway, rather than in .hh
class definitions.  Got caught in a dependency
loop because function contained #define values
but by the time that .hh had been referred to,
they hadn't been defined.

Pulled that function out of the class def
and into an appropriate .cc file, and added
that object to the "test" objects list, and
now builds and links again.
